Rename: Rename and Relocate post components
